NO-TICKET: Refactor into component
