+++ apps/ang-microfront/src/app/app.module.ts
@@ -3,12 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { AppComponent } from './app.component';
 import { appRoutes } from './app.routes';
-import { NxWelcomeComponent } from './page/main/main.components';
+import { NxWelcomeComponent } from './page/main/main.component';
 
 @NgModule({
-  declarations: [AppComponent, NxWelcomeComponent],
+  declarations: [AppComponent ],
   imports: [
     BrowserModule,
+    NxWelcomeComponent,
     RouterModule.forRoot(appRoutes, { initialNavigation: 'enabledBlocking' }),
   ],
   providers: [],
