+++ apps/ang-microfront/src/app/page/main/components/hero/hero.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'org-hero-detail',
+  templateUrl: './hero.component.html',
+  standalone: true
+})
+export class HeroComponent {}
