+++ apps/ang-microfront/src/app/page/main/main.component.html
@@ -1,0 +1,19 @@
+<div class="wrapper">
+      <div class="container">
+        <!--  WELCOME  -->
+        <div id="welcome">
+          <h1>
+            <span> Hello there, </span>
+            Welcome ang-microfront 👋
+          </h1>
+        </div>
+        <!--  HERO  -->
+        <org-hero-detail />
+        <!--  MIDDLE CONTENT  -->
+        <div id="middle-content">
+
+
+        </div>
+
+      </div>
+    </div>
